requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
chromadb>=0.4.22
google-genai>=1.0.0
rank-bm25>=0.2.2
google-api-python-client>=2.111.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
numpy>=1.26.0
tiktoken>=0.5.0
pytest>=7.0
pytest-asyncio>=0.23.0
